fix: logo
