+++ src/renderer/src/components/header.tsx
@@ -8,18 +8,23 @@ import {
 import { APP_NAME } from "@renderer/lib/constants"
 import { Link } from "react-router-dom"
 
+import { Logo } from "./icons/logo"
 import { UserButton } from "./user-button"
 
-const items = [{
-  name: "Home",
-  link: "/",
-}, {
-  name: "Download",
-  link: "/download",
-}, {
-  name: "Blog",
-  link: "/blog",
-}]
+const items = [
+  {
+    name: "Home",
+    link: "/",
+  },
+  {
+    name: "Download",
+    link: "/download",
+  },
+  {
+    name: "Blog",
+    link: "/blog",
+  },
+]
 
 export function Header() {
   // TODO optimize ui in mobile view
@@ -28,7 +33,7 @@ export function Header() {
       <div className="mx-auto flex w-full max-w-[1400px] items-center justify-between">
         <div className="flex items-center gap-8">
           <div className="flex items-center gap-2 text-xl font-bold">
-            <img src="../icon.svg" alt="logo" className="size-8" />
+            <Logo className="size-8" />
             {APP_NAME}
           </div>
           <NavigationMenu className="hidden lg:block">
@@ -36,7 +41,9 @@ export function Header() {
               {items.map((item) => (
                 <NavigationMenuItem key={item.link}>
                   <Link to={item.link}>
-                    <NavigationMenuLink className={navigationMenuTriggerStyle()}>
+                    <NavigationMenuLink
+                      className={navigationMenuTriggerStyle()}
+                    >
                       {item.name}
                     </NavigationMenuLink>
                   </Link>
