+++ src/renderer/src/pages/settings/layout.tsx
@@ -1,3 +1,4 @@
+import { Logo } from "@renderer/components/icons/logo"
 import { Vibrancy } from "@renderer/components/ui/background"
 import { settingTabs } from "@renderer/lib/constants"
 import { SettingsTitle } from "@renderer/modules/settings/title"
@@ -12,7 +13,8 @@ export function Component() {
       <div className="flex flex-1">
         <Vibrancy className="w-44 border-r px-2.5 py-3 pt-3.5">
           <div className="mb-3 flex justify-end">
-            <img src="./icon.svg" alt="logo" className="size-6" />
+
+            <Logo className="size-6" />
           </div>
           {settingTabs.map((t) => (
             <Link
